Rename `module.ts` to `emscripten.ts`
